+++ src/components/icons/BoardViewIcon.tsx
@@ -1,8 +1,6 @@
-import React from "react";
+import type { FC, SVGProps } from "react";
 
-export const BoardViewIcon: React.FC<React.SVGProps<SVGSVGElement>> = (
-  props,
-) => (
+export const BoardViewIcon: FC<SVGProps<SVGSVGElement>> = (props) => (
   <svg
     xmlns="http://www.w3.org/2000/svg"
     viewBox="0 -960 960 960"