+++ src/components/icons/CloseIcon.tsx
@@ -1,6 +1,6 @@
-import React from "react";
+import type { FC, SVGProps } from "react";
 
-export const CloseIcon: React.FC<React.SVGProps<SVGSVGElement>> = (props) => (
+export const CloseIcon: FC<SVGProps<SVGSVGElement>> = (props) => (
   <svg
     xmlns="http://www.w3.org/2000/svg"
     viewBox="0 -960 960 960"