+++ src/components/icons/DensityMediumIcon.tsx
@@ -1,8 +1,6 @@
-import React from "react";
+import type { FC, SVGProps } from "react";
 
-export const DensityMediumIcon: React.FC<React.SVGProps<SVGSVGElement>> = (
-  props,
-) => (
+export const DensityMediumIcon: FC<SVGProps<SVGSVGElement>> = (props) => (
   <svg
     xmlns="http://www.w3.org/2000/svg"
     height="20"